Avoid comparing different integer types

+++ core/src/Xios.cpp
@@ -1411,7 +1411,7 @@ std::vector<std::string> Xios::fileGetFieldIds(const std::string fileId)
 {
     std::vector<xios::CField*> fields = getFile(fileId)->getAllFields();
     std::vector<std::string> fieldIds(fields.size());
-    for (int i = 0; i < fields.size(); i++) {
+    for (size_t i = 0; i < fields.size(); i++) {
         fieldIds[i] = fields[i]->getId();
     }
     return fieldIds;
